Deleted 'main.java' and replaced & renamed it with UserInterface
