--- Main.java
@@ -1,375 +1,0 @@
-import java.io.File;
-import java.io.IOException;
-import java.util.Formatter;
-import java.util.Scanner;
-
-public class Main {
-
-	public static void main(String[] args) {
-
-		/* While loop to keep the program running */
-		System.out.println("Welcome to the Poised Project Registration Interface!\n");
-		boolean runningProgram = true;
-		while (runningProgram) {
-			mainMenu();
-			Scanner userMainMenuInput = new Scanner(System.in);
-			int userMainMenuChoice = userMainMenuInput.nextInt();
-			/* This section deals with the creation of the project object */
-			if (userMainMenuChoice == 1) {
-
-				System.out.println("To create a new project file, please enter the following details: \n");
-				String projectName = projectNameCapture();
-				int projectNumber = projectNumberCapture();
-				String buildingType = buildingTypeCapture();
-				String physicalAddressOfProject = physicalAddressCapture();
-				int erfNumberForProject = erfNumberCapture();
-				float projectFee = projectFeeCapture();
-				float outstandingBalance = outstandingBalance();
-				String projectDeadline = projectDeadline();
-				Project newProjectObject = new Project(projectName, projectNumber, buildingType,
-						physicalAddressOfProject, erfNumberForProject, projectFee, outstandingBalance, projectDeadline);
-				System.out.println("Your project has been successfully registered");
-				System.out.println(newProjectObject.displayProj() + "\n");
-
-				/* This piece of code deals with changing the deadline date */
-				System.out.println("This is your current deadline date: " + " " + newProjectObject.getProjectDeadline()
-				+ " " + "Would you like to change it? 1 - Yes or 2 - No\\n");
-				Scanner userInputToChangeDeadline = new Scanner(System.in);
-				int deadlineChoice = userInputToChangeDeadline.nextInt();
-				if (deadlineChoice == 1) {
-					deadlineChange(newProjectObject);
-				}
-
-				/*This piece of code deals with the change of the current balance*/ 
-				System.out.println(
-						"This is the current balance still to be paid: " + " " + newProjectObject.getCurrBalance() + " "
-								+ "Would you like to change it? 1 - Yes or 2 - No\\n");
-				Scanner userInputToChangeOutstandingBalance = new Scanner(System.in);
-				int choiceToChangeOutstandingBalance = userInputToChangeOutstandingBalance.nextInt();
-				if (choiceToChangeOutstandingBalance == 1) {
-					outstandingBalanceChange(newProjectObject);
-				}
-
-				/* Finalizing the project */
-				System.out.println(
-						"Finalize the project by generating an invoice for the client. 1 - Proceed\n 2 - Return to main menu");
-				Scanner userFinalisationInput = new Scanner(System.in);
-				int userFinalisationChoice = userFinalisationInput.nextInt();
-				if (userFinalisationChoice == 1) {
-
-					System.out.println("An invoice will be generated by the system. Please enter the client's details: \n");
-
-					/* Capturing client details to create client object */
-					Client newClientObject = clientObjectCreation();
-
-					/* Invoice generation */
-					float amountToPay = newProjectObject.getProjectFee() - newProjectObject.getCurrBalance();
-					if (amountToPay > 0) {
-						invoiceGeneration(newClientObject, amountToPay);
-					}
-
-					/* File Creation & input into the file */
-					String fileName = "CompletedProjects.txt";
-					writeToFile(projectName, projectNumber, buildingType, physicalAddressOfProject, erfNumberForProject,
-							projectFee, outstandingBalance, projectDeadline, fileName);
-
-				} else if (userMainMenuChoice == 2) {
-					/* This section will deal with the creation of the contractor object */
-					Contractor newContractor = contractorObjectCreation();
-					System.out.println(newContractor.displayPerson());
-
-					/* Updating the contractor's telephone number */
-					System.out.println("This is the telephone number of contractor: " + " "
-							+ newContractor.getTelephoneNumber() + " .Would you like to update it? 1 - Yes\n 2 - No");
-					
-					Scanner contractorTelephoneNumChange = new Scanner(System.in);
-					int contractorChoiceTelephone = contractorTelephoneNumChange.nextInt();
-					if (contractorChoiceTelephone == 1) {
-						contractorTelephoneChange(newContractor);
-					}
-
-					/* Updating the contractors email address */
-					System.out.println("This is the email address of contractor: " + " "
-							+ newContractor.getEmailAddress() + " . Would you like to update it? 1 - Yes\n 2 - No");
-					
-					Scanner contractorEmailAddressChange = new Scanner(System.in);
-					int contractorChoiceEmail = contractorEmailAddressChange.nextInt();
-					if (contractorChoiceEmail == 1) {
-						contractorEmailAddressChange(newContractor);
-					}
-				}
-			} else if (userMainMenuChoice == 3) {
-				/* This section of code deals with reading the file contents */
-				System.out.println("The projects captured into the system: \n");
-				fileReaderMethod();
-
-			}
-
-		}
-
-	}
-
-	public static void deadlineChange(Project newProjectObject) {
-		System.out.println("Enter the new date: ");
-		Scanner newDateEntry = new Scanner(System.in);
-		String newDate = newDateEntry.nextLine();
-		newProjectObject.setProjectDeadline(newDate);
-		System.out.println("Your new deadline has been successfully registered");
-		System.out.println(newProjectObject.displayProj() + "\n");
-	}
-
-	public static void outstandingBalanceChange(Project newProjectObject) {
-		System.out.println("Enter the new balance: ");
-		Scanner newBalanceInput = new Scanner(System.in);
-		float newBalance = newBalanceInput.nextFloat();
-		newProjectObject.setCurrBalance(newBalance);
-	}
-
-	public static void fileReaderMethod() {
-		try {
-			File savedProjects = new File("CompletedProjects.txt");
-			Scanner fileReader = new Scanner(savedProjects);
-			while (fileReader.hasNext()) {
-				System.out.println(fileReader.nextLine() + "\n");
-			}
-			fileReader.close();
-		} catch (IOException e) {
-			System.out.println("An error has occured in attempting to read he file");
-		}
-	}
-
-	public static void contractorEmailAddressChange(Contractor newContractor) {
-		System.out.println("Enter the new email address:");
-		Scanner newEmailAddressInput = new Scanner(System.in);
-		String newEmailAddress = newEmailAddressInput.nextLine();
-		newContractor.setEmailAddress(newEmailAddress);
-		System.out.println("New email address registered!\n");
-		System.out.println(newContractor.displayPerson());
-	}
-
-	public static void contractorTelephoneChange(Contractor newContractor) {
-		System.out.println("Enter the new telephone number:");
-		Scanner newTelephoneNumberInput = new Scanner(System.in);
-		int newTelephoneNumber = newTelephoneNumberInput.nextInt();
-		newContractor.setTelephoneNumber(newTelephoneNumber);
-		System.out.println("New telephone number registered!\n");
-		System.out.println(newContractor.displayPerson());
-	}
-
-	public static Contractor contractorObjectCreation() {
-		System.out.println("Please enter the following details to register a new contractor: \n");
-		String nameOfContractor = contractorName();
-		int telephoneNumberOfContractor = contractorTelephoneNumber();
-		String emailAddressOfContractor = contractorEmailAddress();
-		String physicalAddressOfContractor = contractorPhysicalAddress();
-		Contractor newContractor = new Contractor(nameOfContractor, telephoneNumberOfContractor,
-				emailAddressOfContractor, physicalAddressOfContractor);
-		System.out.println("New contractor registered!\n");
-		return newContractor;
-	}
-
-	public static void writeToFile(String projectName, int projectNumber, String buildingType,
-			String physicalAddressOfProject, int erfNumberForProject, float projectFee, float outstandingBalance,
-			String projectDeadline, String fileName) {
-		try {
-			Formatter outputFile = new Formatter(fileName);
-			outputFile.format(
-					"Project Name: %s%nProject Number: %d%nBuilding Type: %s%nPhysical Address: %s%n"
-							+ "ERF Number: %d%nProject Fee: %f%nOutstanding Balance: %f%nProject Deadline: %s%n",
-							projectName, projectNumber, buildingType, physicalAddressOfProject, erfNumberForProject,
-							projectFee, outstandingBalance, projectDeadline);
-			outputFile.close();
-		} catch (IOException e) {
-			System.out.println("An error has occured writing to the file" + e);
-		}
-	}
-
-	public static void invoiceGeneration(Client newClientObject, float amountToPay) {
-		int invoiceNumber = (int) (Math.random() * 1000);
-		System.out.println("Invoice " + " " + invoiceNumber + "\n");
-		System.out.println(newClientObject.displayPerson() + "\n");
-		System.out.println("Outstanding balance: " + " " + amountToPay);
-	}
-
-	public static Client clientObjectCreation() {
-		String nameOfClient = customerName();
-		int telephoneOfClient = customerTelephoneNumber();
-		String emailAddressOfClient = customerEmailAddress();
-		String physicalAddressOfClient = customerPhysicalAddress();
-		Client newClientObject = new Client(nameOfClient, telephoneOfClient, emailAddressOfClient,
-				physicalAddressOfClient);
-		
-		return newClientObject;
-	}
-
-	// Methods to capture the details of the project
-	public static String projectNameCapture() {
-		System.out.println("The name of the project: ");// Name of the project
-			Scanner inputForName = new Scanner(System.in);
-			String nameForProject = inputForName.nextLine();
-		
-			return nameForProject;
-		
-		
-		
-
-		
-	}
-
-	public static int projectNumberCapture() {
-		System.out.println("The project number: ");
-		Scanner inputForNum = new Scanner(System.in);
-		int numForProject = inputForNum.nextInt();
-
-		return numForProject;
-	}
-
-	public static String buildingTypeCapture() {
-		System.out.println("Building type (eg. Apartment, House etc.): ");
-		Scanner inputForBuild = new Scanner(System.in);
-		String buildingType = inputForBuild.nextLine();
-
-		return buildingType;
-	}
-
-	public static String physicalAddressCapture() {
-		System.out.println("The physical address of the project: ");
-		Scanner inputForPhysAdd = new Scanner(System.in);
-		String physicalAddress = inputForPhysAdd.nextLine();
-
-		return physicalAddress;
-	}
-
-	public static int erfNumberCapture() {
-		System.out.println("The ERF number: ");
-		Scanner inputForErf = new Scanner(System.in);
-		int erf = inputForErf.nextInt();
-
-		return erf;
-	}
-
-	public static float projectFeeCapture() {
-		System.out.println("The total fee for the project: R ");
-		Scanner inputForProjFee = new Scanner(System.in);
-		float fees = inputForProjFee.nextFloat();
-
-		return Math.round(fees * 100) / 100;
-	}
-
-	public static float outstandingBalance() {
-		System.out.println("The fees still outstanding: R ");
-		Scanner inputForProjBal = new Scanner(System.in);
-		float balance = inputForProjBal.nextFloat();
-
-		return Math.round(balance * 100) / 100;
-	}
-
-	public static String projectDeadline() { 
-		System.out.println("The deadline for the project: " + "\n");
-		Scanner inputForDeadline = new Scanner(System.in);
-		String deadline = inputForDeadline.nextLine();
-
-		return deadline;
-	}
-
-	// Methods for capturing the details of the contractors
-	public static String contractorName() {
-		System.out.println("Contractor's name: \n");
-		Scanner contrName = new Scanner(System.in);
-		String conName = contrName.nextLine();
-
-		return conName;
-	}
-
-	public static int contractorTelephoneNumber() {
-		System.out.println("Contractor's telephone number: \n");
-		Scanner contrTele = new Scanner(System.in);
-		int conTele = contrTele.nextInt();
-
-		return conTele;
-	}
-
-	public static String contractorEmailAddress() {
-		System.out.println("Contractor's email address: \n");
-		Scanner contrEmail = new Scanner(System.in);
-		String conEmail = contrEmail.nextLine();
-
-		return conEmail;
-	}
-
-	public static String contractorPhysicalAddress() {
-		System.out.println("Contractor's physical address: ");
-		Scanner contrPhysAdd = new Scanner(System.in);
-		String conPhysAdd = contrPhysAdd.nextLine();
-
-		return conPhysAdd;
-	}
-
-	// Methods for capturing the details of the customers
-	public static String customerName() {
-		System.out.println("Customer's name: \n");
-		Scanner clientName = new Scanner(System.in);
-		String cliName = clientName.nextLine();
-
-		return cliName;
-	}
-
-	public static int customerTelephoneNumber() {
-		System.out.println("Client's telephone number: \n");
-		Scanner clientTele = new Scanner(System.in);
-		int cliTele = clientTele.nextInt();
-
-		return cliTele;
-	}
-
-	public static String customerEmailAddress() {
-		System.out.println("Client's email address: \n");
-		Scanner clientEmail = new Scanner(System.in);
-		String cliEmail = clientEmail.nextLine();
-
-		return cliEmail;
-	}
-
-	public static String customerPhysicalAddress() {
-		System.out.println("Client's physical address: ");
-		Scanner clientPhysAdd = new Scanner(System.in);
-		String cliPhysAdd = clientPhysAdd.nextLine();
-
-		return cliPhysAdd;
-	}
-
-	// Methods for capturing the details of the architect
-	public static String architectName() {
-		Scanner archName = new Scanner(System.in);
-		String arcName = archName.nextLine();
-
-		return arcName;
-	}
-
-	public static int architectTelNum() {
-		Scanner archTele = new Scanner(System.in);
-		int arcTele = archTele.nextInt();
-
-		return arcTele;
-	}
-
-	public static String architectEmailAddress() {
-		Scanner archEmail = new Scanner(System.in);
-		String arcEmail = archEmail.nextLine();
-
-		return arcEmail;
-	}
-
-	public static String architectPhysAdd() {
-		Scanner archPhysAdd = new Scanner(System.in);
-		String arcPhysAdd = archPhysAdd.nextLine();
-
-		return arcPhysAdd;
-	}
-
-	public static void mainMenu() {
-		System.out.println("Please select one of the following options: \n");
-		System.out.println(" 1. Project registration\n" + " 2. Contractor registration\n" + " 3. View all projects\n");
-	}
-
-}
